+++ Dracula/ExternalAPIs/Kokoro/CardRenderHook.cs
@@ -11,7 +11,6 @@ public partial interface IKokoroApi
 
 public interface ICardRenderHook
 {
-	Vec ModifyTextCardScale(G g, Card card) => Vec.One;
 	Matrix ModifyNonTextCardRenderMatrix(G g, Card card, List<CardAction> actions) => Matrix.Identity;
 	Matrix ModifyCardActionRenderMatrix(G g, Card card, List<CardAction> actions, CardAction action, int actionWidth) => Matrix.Identity;
 }