+++ Dracula/GameEnums.cs
@@ -4,3 +4,6 @@ namespace Shockah.Dracula;
 
 [EnumByName(typeof(Spr))]
 internal static partial class StableSpr { }
+
+[EnumByName(typeof(UK))]
+internal static partial class StableUK { }